bom_comparison_tool/requirements.txt
openpyxl
python-docx
PyPDF2
pytest
